abstracted components and resolved pr issues
