+++ src/components/Pair2/DeleteButton.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import { Button } from 'antd'
+
+type Props = {
+  onClick?: React.MouseEventHandler
+}
+const ButtonStyles = { margin: '5px' }
+
+export const DeleteButton = (props: Props) => {
+  return (
+    <Button danger onClick={props.onClick} style={ButtonStyles}>
+      Delete
+    </Button>
+  )
+}
+
+export default DeleteButton
